More feedback

+++ internal/ludwig/type.go
@@ -374,7 +374,7 @@ type FrameObject struct {
 	ScrDotLine    int
 	Span          *SpanObject
 	ReturnFrame   *FrameObject
-	InputCount    uint32
+	InputCount    uint64
 	SpaceLimit    int
 	SpaceLeft     int
 	TextModified  bool
